PR feedback addressed
