+++ src/NuGetPackageVerifier/Rules/AssemblyHasCommitHashAttributeRule.cs
@@ -1,4 +1,4 @@
-// Copyright (c) .NET Foundation. All rights reserved.
+﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System;
@@ -15,8 +15,7 @@ namespace NuGetPackageVerifier.Rules
     {
         public IEnumerable<PackageVerifierIssue> Validate(PackageAnalysisContext context)
         {
-            AssemblyHasAttributeHelper.GetAssemblyAttributesData(context);
-
+            AssemblyAttributesDataHelper.SetAssemblyAttributesData(context);
             foreach (var assemblyData in context.AssemblyData)
             {
                 var fileName = Path.GetFileNameWithoutExtension(assemblyData.Key);
