+++ src/NuGetPackageVerifier/Rules/AssemblyHasCopyrightAttributeRule.cs
@@ -1,4 +1,4 @@
-// Copyright (c) .NET Foundation. All rights reserved.
+﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System.Collections.Generic;
@@ -13,7 +13,7 @@ namespace NuGetPackageVerifier.Rules
     {
         public IEnumerable<PackageVerifierIssue> Validate(PackageAnalysisContext context)
         {
-            AssemblyHasAttributeHelper.GetAssemblyAttributesData(context);
+            AssemblyAttributesDataHelper.SetAssemblyAttributesData(context);
             foreach (var assemblyData in context.AssemblyData)
             {
                 if (!HasCopyrightAttribute(assemblyData.Value.AssemblyAttributes))
