+++ src/NuGetPackageVerifier/Rules/AssemblyHasNeutralResourcesLanguageAttributeRule.cs
@@ -1,4 +1,4 @@
-// Copyright (c) .NET Foundation. All rights reserved.
+﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System;
@@ -14,7 +14,7 @@ namespace NuGetPackageVerifier.Rules
     {
         public IEnumerable<PackageVerifierIssue> Validate(PackageAnalysisContext context)
         {
-            AssemblyHasAttributeHelper.GetAssemblyAttributesData(context);
+            AssemblyAttributesDataHelper.SetAssemblyAttributesData(context);
             foreach (var assemblyData in context.AssemblyData)
             {
                 if (!HasNeutralResourcesLanguageAttribute(assemblyData.Value.AssemblyAttributes))
